LayoutTests/editing/execCommand/switch-list-type-with-inner-list-expected.txt

This tests switching an unordered list with a nested list to an ordered list.
| "
"
| <ol>
|   <li>
|     "<#selection-caret>hello"
|   "
    "
|   <ol>
|     "
    "
|     <li>
|       "world"
|     "
    "
|   "
    "
|   <li>
|     "webkit"
| "
"
